Refactor pellet settings
remove inner classes, more pointless methods
